Add create2deployer to genesis

+++ op-chain-ops/genesis/layer_two.go
@@ -61,7 +61,7 @@ func BuildL2Genesis(config *DeployConfig, l1StartBlock *types.Block) (*core.Gene
 			}
 			db.CreateAccount(codeAddr)
 			db.SetState(addr, ImplementationSlot, codeAddr.Hash())
-		} else {
+		} else if db.Exist(addr) {
 			db.DeleteState(addr, AdminSlot)
 		}
 		if err := setupPredeploy(db, deployResults, storage, name, addr, codeAddr); err != nil {
